Remove _solutionOpenedRaised
The project added solution handler was being skipped when the solution
opened raised bool was set the false, and with recent dev17 changes,
none of the code paths that previously triggered the bool being set to
true are being hit any longer. Therefore, new projects are not being
added into the project system cache, and other APIs that expect the
project to be in the project system cache are therefore failing.

The comment on the variable say it's to ignore project renamed events
before the solution is fully open, but it doesn't expain in what
scenario a project would be renamed on solution load.